+++ scilab/modules/types/includes/container.hxx
@@ -1,13 +1,13 @@
 /*
  *  Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
  *  Copyright (C) 2010-2010 - DIGITEO - Bruno JOFRET
- *
+ * 
  *  This file must be used under the terms of the CeCILL.
  *  This source file is licensed as described in the file COPYING, which
  *  you should have received as part of this distribution.  The terms
  *  are also available at
  *  http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
- *
+ * 
  */
 
 #ifndef __CONTAINER_HXX__
@@ -26,7 +26,7 @@ namespace types
 
         virtual void            whoAmI(void) { std::cout << "types::Container"; }
 
-        virtual size_t             size_get() = 0 ;
+        virtual int             size_get() = 0 ;
 
         virtual Container*  getAsContainer(void) { return this; }
