+++ scilab/modules/types/includes/list.hxx
@@ -1,13 +1,13 @@
 /*
  *  Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
  *  Copyright (C) 2010-2010 - DIGITEO - Bruno JOFRET
- *
+ * 
  *  This file must be used under the terms of the CeCILL.
  *  This source file is licensed as described in the file COPYING, which
  *  you should have received as part of this distribution.  The terms
  *  are also available at
  *  http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
- *
+ * 
  */
 
 #ifndef __LIST_HH__
@@ -22,7 +22,7 @@ namespace types
     {
     public :
                                         List();
-                                        ~List();
+                                        ~List(); 
 
     private :
                                         List(List *_oListCopyMe);
@@ -30,11 +30,11 @@ namespace types
         std::vector<InternalType *>*    getData();
 
     public :
-        size_t                             size_get();
+        int                             size_get(); 
 
         void                            whoAmI(void) { std::cout << "types::List"; };
 
-        RealType                        getType(void) { return RealList; }
+        RealType                        getType(void) { return RealList; } 
 
         /**
         ** append(InternalType *_typedValue)
@@ -52,8 +52,8 @@ namespace types
 
         List*                           getAsList(void) { return this; }
 
-        bool                            insert(size_t _iSeqCount, size_t* _piSeqCoord, size_t* _piMaxDim, vector<types::InternalType*>* _poSource, bool _bAsVector);
-        std::vector<InternalType*>      extract(size_t _iSeqCount, size_t* _piSeqCoord, size_t* _piMaxDim, size_t* _piDimSize, bool _bAsVector);
+        bool                            insert(int _iSeqCount, int* _piSeqCoord, int* _piMaxDim, vector<types::InternalType*>* _poSource, bool _bAsVector);
+        std::vector<InternalType*>      extract(int _iSeqCount, int* _piSeqCoord, int* _piMaxDim, int* _piDimSize, bool _bAsVector);
 
         /* return type as string ( double, int, cell, list, ... )*/
         virtual wstring                 getTypeStr() {return L"list";}
